Update SettingUpDevelopmentEnvironment.md
After some wonderful help, I'm updating the setup with what I've learned and how we fixed my multiple build issues

I'm not familiar with the ML this page uses, so I've stuck the major points as bullets. I wanted to place them as subsections, but this works too.

It's pretty extensive but it works. I hope it'll help people get this working too.