Fix Waffle Demo Parent RelativePath
Relative Path needed to point further up the tree for waffle parent.
This was not clear initially as an issue due to prior builds already
being in place.  To reproduce, the local waffle repo entries were
deleted and the build was attempted via command prompt.  That resulted
in the failure seen by Daniel.  Now that things are stable this issue
should subside :)