Added coverity
added coverity for more indepth scanning as part of travis ci.  To use
this, it requires setup over on coverity.  Currently it is coded to my
location which should be fine for the time being as it also requires to
be run on special branch that needs setup for it to actually do
anything.