Mocking Tomcat 8
Change all test cases to make sue of mocking thus removing all but
needed items for waffle.  This makes intent clearer and might ultimately
lead to a solution of test cases that can apply to all versions.  At the
very least, it decreases code quite a bit that is essentially no more
than noise.  Uses mockito and powermock as needed.