Drop legacy deprecated base64, we use guava
As indicated in release 1.7, when 1.8 was to be worked the legacy base64
code would be removed.  Technically this should have been done right out
the gate since the unit test was changed to test guava rather than
remaining with the deprecated code to be tested.  Anyway, this is now
gone.