Testing WARS start...
Mockito is progressing forwards fast.  This has caused changes on
internal package changes which various projects improperly used internal
pieces and thus are now broke.  In light of this, I've decided to move
over to jmockit in an effort to get rid of powermock whitefox in favor
of deencapsulation.  In fact, this is cleaner!  Additionally, the
doThrow powermock is use is now available directly in mockito.  I still
like mockito so that will remain for now.

What I haven't touched is catch-exception usage.  Leaving that at prior
mockito level and will address that project directly.  Since
catch-exception relies on mockito we won't see it go anywhere anytime
soon.