Javadoc Issue Resolved - Jetty Dummy Class
Removed skip javadoc in jetty pom.  It was breaking others.  Javadocs no
longer create empty javadocs thus the reason that was there.  That in
turn caused multi module build to break at various points due to
threading.  In order to resolve this and not get empty javadoc, added
dummy class which allows us to get everything else we want.