Merge maven site and readme markdown for site
update antrun plugin to 1.8
update surefire plugin to 2.18.1
update jgit to 3.6.0 release.

In order to replace gh-pages site from original waffle version, merge
the readme markdown into the maven generated twitter bootstrap page
layout.  This is actually a large improvement over the original as
readme has changed quite a bit and had lots more useful information to
provide.  At the moment this does mean some duplication with the readme
that will be addressed later.

This results in doxia-module-markdown and wagon-git being added.  There
is not presently a good solid way to do this so write-up will be coming.
As always multi-module issues persist thus multiple ways around this.
After much work on this, this seems best solution for now but requires a
little setup to work.

mvn site to build the site
mvn site:deploy to deply the site.  requires setup in .m2 for this to
work which is dependent on user settings.