Replace catch-exception with fluent-exception-rule
catch-exception relied on older mockito version.  In newer mockito
version which we use, code has been make class default thus no longer
available for external use.  I worked on catch-exception library trying
to address this but that never panned out as it was rather difficult
thing to address.  That said, this new library 'fluent-exception-rule'
utilizes junit rules and only relies on assertj which we also use.  In
fact, so did catch-exception.  The overall code is even cleaner to read.