[issue #203] Catch Win32Exception and make 401
A fix to capture tomcat 6, 7, and 8 negotiate authenticator
Win32Exception which started occurring around June 2015 per issue #107.
Around that time Exception was switched out for IOException which
allowed Win32Exception to occur.  Now catch that and perform same
actions as IOException.  Left original unit tests capturing fall through
and added new check as this shows Negotiate, NTLM where the comma was
normally expected to be a space.  This does fix the condition but I'm
not 100% certain this is correct from unit test side.