Fix DelegatingNegotiateSecurityFilter use of custom auth
For the SpringSecurity 4 and 5 modules DelegatingNegotiateSecurityFilter, if both a custom AuthenticationManager and a custom AuthenticationSuccessHandler are specified, the custom Authentication is added to the SecurityContext, but is not passed to the custom success handler.  This means the context has a different Authentication than the success handler.

This is fixed for both spring security modules.

Issues:
https://github.com/Waffle/waffle/issues/453