[jdk] Stop using deprecated interface and just directly configure group
group has been deprecated for years, it's deleted from newer jdks.  Since we only use that interface but otherwise provide the implementation, simply stop using it and continue as is.  This is untested as far as jdk 14/15 goes but all tests do pass.  Additional testing will be confirmed prior to release.