[.net] Remove build of java code as releases are separate and projects will be split later
eventually will split this out as .net side rarely ever changes and is not most used (or at least known to be used).  In meantime, just turn off .net java build but allow other cleanup to occur.  This is to allow builds to be restored on github actions as it was run on appvoyer but that hasn't run since 2017 likely due to some security changes as no longer linked or needed (maven is way too old on that build).