Marked GroupPrincipal as deprecated
The GroupPrincipal does not conform to JAAS where all principles should be in the Subject.Principal set.  Hiding group names or role names in a "GroupPrincipal" is not JAAS compliant and the PR which created this incorrect class also broke the WindowsLoginModule for all JAAS compliant users, such as Tomcat.