Modified The WindowsLoginModule Tests to pass only if correct RolePrincipals are present, but regardless of whether or not the incorrect "Roles" GroupPrincipal is present
This is done to allow deprecated code to be present or not when testing.  It does not do an exact match on which principles to expect, as it allows unspecified GroupPrincipals to be present.  The tests will pass if the correct information is contained in the Subject.Principals, but will not fail if additional Principals are defined.