Eek/fix/windows login module (#1116)
* Removed the incorrect GroupPrincipal "Roles" from the code

The existing code for WindowsLoginModule (since October 2016) does not bind the Windows (or Active DIrectory) group (names) to the Subject. It binds an object named "Roles" to the Subject. It therefore does not fulfill the basic function required of a JAAS WindowsLoginModule.  Refer to Issue #1114.

The packaging into a GroupPrincipal (which did not exist before the breaking change), was made to get it to work with WildFly (earlier versions as it does not work with latest versions in any case). Every Windows Group to which the User belongs, was placed into a new list of Principals (named "Roles") which was then packaged as a Principal (the GroupPrincipal) and added to the Subject.Principals list. This is just plain wrong, as it assumes that any user of the subject will know about the custom "Roles", and be able to extract the Principals from this list, This assumption is obviously invalid. The JAAS specified way is to obtain it from the Subject.Principles.getName(). The code modification also removed the ability to obtain the Windows Group (or JAAS roles) from the Subject.Principles.getName() as specified by JAAS.

* Fixed the tests to exclude the incorrect GroupPrinciple named "Roles"

* Update WindowsLoginModuleTest.java

* Corrected the tests to accept Windows groups as (Role)Principles

Refer to issue #1114.

* Fixed the changed name in the assert.

* Fixed the failing tests

* Delete GroupPrincipal.java

GroupPrincipal breaks the JAAS Principal and should be removed.

* Delete the GroupPrincipal tests as as the GroupPrincipal was deleted

* Removed the incorrect GroupPrincipal "Roles" from the code

The existing code for WindowsLoginModule (since October 2016) does not bind the Windows (or Active DIrectory) group (names) to the Subject. It binds an object named "Roles" to the Subject. It therefore does not fulfill the basic function required of a JAAS WindowsLoginModule.  Refer to Issue #1114.

The packaging into a GroupPrincipal (which did not exist before the breaking change), was made to get it to work with WildFly (earlier versions as it does not work with latest versions in any case). Every Windows Group to which the User belongs, was placed into a new list of Principals (named "Roles") which was then packaged as a Principal (the GroupPrincipal) and added to the Subject.Principals list. This is just plain wrong, as it assumes that any user of the subject will know about the custom "Roles", and be able to extract the Principals from this list, This assumption is obviously invalid. The JAAS specified way is to obtain it from the Subject.Principles.getName(). The code modification also removed the ability to obtain the Windows Group (or JAAS roles) from the Subject.Principles.getName() as specified by JAAS.

* Fixed the tests to exclude the incorrect GroupPrinciple named "Roles"

* Removed the warnings on specific desrialization tests.

* Disabled dserialization ban on serializable test.

* Suppressed warnings when testing specifically for Serializability.

Co-authored-by: Frederick Peter Eek <48279754+FreekEek@users.noreply.github.com>