[ci] Add readme's that partially got cargo working with todo on them
will try to work to green later, this is a start and shouldn't affect the old way it was done.

For jaas, needs logon form
For mixed and negotiate, needs to handle caffeine better as service loader is a problem and cargo needs it defined in both places to provide to tomcat.